Extract ShadowJarPublication to its own configurer
